Remove unused RedHat-specific systemd EnvironmentFile
Without this change, Puppet attempts to create a directory at
`/etc/sysconfig/nomad` when init_style is systemd.  This is a problem
because recent versions of Ubuntu and Debian use systemd but do not have an
`/etc/sysconfig` directory (they use `/etc/default` instead), thus puppet
cannot create the nomad directory.

As far as I can tell, there's not reason to use this directory so it's
simply been removed rather than reworking it to support the equivalent in
Ubuntu and Debian.